win-build: iconv updated and executable code removed from header
