+++ src/lib_ccx/ccx_encoders_common.h
@@ -2,7 +2,7 @@
 #define _CC_ENCODER_COMMON_H
 
 #ifdef WIN32
-	#include "..\\win_iconv\\win_iconv.h"
+	#include "..\\win_iconv\\iconv.h"
 #else
 	#include "iconv.h"
 #endif
